moved tmp_dir fixtures to fixtues.py; moved testfile to appropiate location; moved the functions out of the class
